@@ -35,14 +35,14 @@
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/animations": "13.3.10",
+    "@angular/common": "13.3.10",
+    "@angular/compiler": "13.3.10",
+    "@angular/core": "13.3.10",
+    "@angular/forms": "13.3.10",
+    "@angular/platform-browser": "13.3.10",
+    "@angular/platform-browser-dynamic": "13.3.10",
+    "@angular/router": "13.3.10",
     "@nativescript/core": "~8.1.0",
     "@nativescript/theme": "~3.0.0",
     "@ngx-translate/core": "~14.0.0",
@@ -51,29 +51,29 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.2.2",
+    "@angular-devkit/build-angular": "13.3.7",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
     "@angular-eslint/template-parser": "13.0.1",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular/compiler-cli": "13.3.10",
     "@jsdevtools/coverage-istanbul-loader": "3.0.5",
-    "@nativescript/types": "~8.1.0",
+    "@nativescript/types": "~8.2.0",
     "@nativescript/unit-test-runner": "^3.0.2",
     "@nativescript/webpack": "~5.0.0",
-    "@ngtools/webpack": "13.2.2",
-    "@nrwl/angular": "13.7.3",
-    "@nrwl/cli": "13.7.3",
-    "@nrwl/eslint-plugin-nx": "13.7.3",
-    "@nrwl/jest": "13.7.3",
-    "@nrwl/node": "13.7.3",
-    "@nrwl/tao": "13.7.3",
-    "@nrwl/workspace": "13.7.3",
-    "@nstudio/angular": "13.4.3",
-    "@nstudio/nativescript": "13.4.3",
-    "@nstudio/nativescript-angular": "13.4.3",
-    "@nstudio/xplat": "13.4.3",
+    "@ngtools/webpack": "13.3.7",
+    "@nrwl/angular": "14.1.9",
+    "@nrwl/cli": "14.1.9",
+    "@nrwl/eslint-plugin-nx": "14.1.9",
+    "@nrwl/jest": "14.1.9",
+    "@nrwl/js": "14.1.9",
+    "@nrwl/node": "14.1.9",
+    "@nrwl/workspace": "14.1.9",
+    "@nstudio/angular": "14.0.1",
+    "@nstudio/nativescript": "14.0.1",
+    "@nstudio/nativescript-angular": "14.0.1",
+    "@nstudio/xplat": "14.0.1",
     "@types/jasmine": "^3.7.2",
-    "@types/jest": "27.0.2",
+    "@types/jest": "27.4.1",
     "@types/node": "14.14.33",
     "@types/sinon": "^10.0.2",
     "@typescript-eslint/eslint-plugin": "4.19.0",
@@ -84,27 +84,28 @@
     "eslint-config-prettier": "8.1.0",
     "fs-extra": "^9.0.1",
     "husky": "~7.0.1",
-    "jest": "27.2.3",
-    "karma": "6.3.9",
+    "jest": "27.5.1",
+    "karma": "6.3.20",
     "karma-coverage": "2.0.3",
     "karma-jasmine": "4.0.1",
-    "karma-nativescript-launcher": "0.4.0",
     "karma-mocha-reporter": "^2.2.5",
+    "karma-nativescript-launcher": "0.4.0",
     "karma-sinon": "^1.0.5",
     "lint-staged": "^11.0.0",
-    "ng-packagr": "~13.3.0",
+    "ng-packagr": "13.3.1",
+    "nx": "14.1.9",
     "nyc": "15.1.0",
     "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
     "prettier": "2.5.1",
     "sass": "^1.32.0",
     "sinon": "^11.1.1",
-    "ts-jest": "27.0.5",
+    "ts-jest": "27.1.4",
     "ts-node": "9.1.1",
     "tslib": "^2.1.0",
-    "typescript": "4.5.5"
+    "typescript": "4.6.4"
   },
   "xplat": {
     "prefix": "nativescript",
