@@ -1,3 +1,3 @@
 const { getJestProjects } = require('@nrwl/jest');
 
-module.exports = { projects: getJestProjects() };
+export default { projects: getJestProjects() };
