@@ -1,5 +1,6 @@
-module.exports = {
-  displayName: 'zone-js',
+/* eslint-disable */
+export default {
+  displayName: 'angular',
   preset: '../../jest.preset.js',
   globals: {
     'ts-jest': { tsconfig: '<rootDir>/tsconfig.spec.json' },
@@ -9,5 +10,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/packages/zone-js',
+  coverageDirectory: '../../coverage/packages/angular',
 };
